feat: xcm entity, load config on start
