+++ src/renderer/domain/shared-kernel.ts
@@ -1,10 +1,14 @@
+export type ChainId = HexString;
 export type HexString = `0x${string}`;
+
 export type Address = string;
 export type AccountId = HexString;
+
 export type Threshold = number;
+
 export type CallData = HexString;
 export type CallHash = HexString;
-export type ChainId = HexString;
+
 export type EraIndex = number;
 export type Timepoint = {
   height: number;
