+++ src/renderer/entities/xcm/index.ts
@@ -1,0 +1,1 @@
+export * as xcmModel from './model/xcm-model';
