+++ src/renderer/shared/core/index.ts
@@ -1,0 +1,1 @@
+export * as kernelModel from './model/kernel-model';
