+++ src/renderer/shared/core/model/kernel-model.ts
@@ -1,0 +1,7 @@
+import { createEvent } from 'effector';
+
+const appStarted = createEvent();
+
+export const events = {
+  appStarted,
+};
